Refactor manifest class to separate file
