+++ Packages/Tracking OpenXR/Editor/Scripts/AndroidManifest.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: db9d8db83ab841b6a52a16ceda247cb0
+timeCreated: 1658216771
